erasmus/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-filter>=24.2
psycopg[binary]>=3.2
python-dotenv>=1.0
django-allauth>=65.0